controle_compras/requirements.txt
Django==2.2.14
django-treenode==0.15.0
